// src/types/interfaces/location.ts
/* tslint:disable */

export interface ILocation 
{
    id: string;
    name?: string; 
    addressStreetNumber: string; 
    addressStreetName: string; 
    addressSuburb: string; 
    addressCityTown: string; 
    addressProvinceState: string; 
    addresCountryCode: string; 
    iconName: string; 
}